Standardize lock metadata
Squashed commit of the following:

commit d1aef966ca6c13c67e13d204040afbd0787e62ba
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 15:37:59 2026 -0400

    test fixes

commit 7a14be93b67db922d05f5c923f139066344b4b17
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 15:24:09 2026 -0400

    test fixes

commit 578e829c3ac8c75b2eff8b29ae86ab2d7f8900f4
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 12:53:48 2026 -0400

    ci test updates

commit 34461c25422a508946a8476ab9a1aeee6dc5af3c
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 11:40:58 2026 -0400

    unit test updates

commit ec2d89fd6a299608f54655ef0bdf799491eeaaa8
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 09:47:09 2026 -0400

    increase unit tests

commit fbe53cafcb90bedb65b055d7fa60ff4dd3e651c1
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 08:14:32 2026 -0400

    ci fixes

commit 19919e69ccfd814d0322dc08d0e2f1ab25ef39c5
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 20 07:44:02 2026 -0400

    tidy up lock metadata feature

commit e2f444da466b84150f49c283e440f2517e0fa790
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 13 13:53:28 2026 -0400

    checkpoint

commit 54457620137adfb2780b5e83e0c7da2d55cc1959
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Mon Apr 13 12:40:40 2026 -0400

    checkpoint

+++ src/zxfer_runtime.sh
@@ -994,7 +994,7 @@ zxfer_get_max_yield_iterations() {
 # bootstrap so downstream code sees consistent defaults and runtime state.
 zxfer_init_runtime_metadata() {
 	# zxfer version
-	g_zxfer_version="2.0.0-20260411"
+	g_zxfer_version="2.0.0-20260413"
 }
 
 # Purpose: Initialize the option defaults before later helpers depend on it.
@@ -1100,6 +1100,7 @@ zxfer_init_transport_remote_defaults() {
 	g_zxfer_ssh_control_socket_action_command=""
 	g_zxfer_ssh_control_socket_lock_dir_result=""
 	g_zxfer_ssh_control_socket_lock_error=""
+	g_zxfer_ssh_control_socket_lease_count_result=""
 	g_zxfer_remote_capability_cache_ttl=15
 	g_zxfer_remote_capability_cache_wait_retries=5
 	g_source_operating_system=""
@@ -1138,6 +1139,9 @@ zxfer_init_runtime_state_defaults() {
 	g_zxfer_effective_tmpdir=""
 	g_zxfer_effective_tmpdir_requested=""
 	g_zxfer_temp_file_result=""
+	if command -v zxfer_reset_owned_lock_tracking >/dev/null 2>&1; then
+		zxfer_reset_owned_lock_tracking
+	fi
 	zxfer_reset_runtime_artifact_state
 	g_zxfer_profile_start_epoch=$(date '+%s' 2>/dev/null || :)
 	g_zxfer_profile_has_data=0
@@ -1324,6 +1328,9 @@ zxfer_trap_exit() {
 	if command -v zxfer_close_all_ssh_control_sockets >/dev/null 2>&1; then
 		zxfer_close_all_ssh_control_sockets
 	fi
+	if command -v zxfer_release_registered_owned_locks >/dev/null 2>&1; then
+		zxfer_release_registered_owned_locks || :
+	fi
 
 	zxfer_cleanup_registered_runtime_artifacts
 
@@ -1338,6 +1345,14 @@ zxfer_trap_exit() {
 	if l_tmpdir=$(zxfer_try_get_effective_tmpdir 2>/dev/null); then
 		for l_temp_file in "$l_tmpdir/${g_zxfer_temp_prefix:-zxfer.unset}".*; do
 			[ -e "$l_temp_file" ] || continue
+			if command -v zxfer_remote_host_cache_cleanup_conflicts_with_path >/dev/null 2>&1 &&
+				zxfer_remote_host_cache_cleanup_conflicts_with_path "$l_temp_file"; then
+				continue
+			fi
+			if command -v zxfer_owned_lock_cleanup_conflicts_with_path >/dev/null 2>&1 &&
+				zxfer_owned_lock_cleanup_conflicts_with_path "$l_temp_file"; then
+				continue
+			fi
 			rm -rf "$l_temp_file"
 		done
 	fi
@@ -1349,8 +1364,6 @@ zxfer_trap_exit() {
 	fi
 	if command -v zxfer_cleanup_remote_host_cache_roots >/dev/null 2>&1; then
 		zxfer_cleanup_remote_host_cache_roots >/dev/null 2>&1 || :
-	elif command -v zxfer_cleanup_remote_host_cache_roots_if_empty >/dev/null 2>&1; then
-		zxfer_cleanup_remote_host_cache_roots_if_empty >/dev/null 2>&1 || :
 	fi
 
 	if [ "${g_services_need_relaunch:-0}" -eq 1 ]; then
